Upate README;Adjust file structure;add test for mrcProfiler
