+++ CMakeLists.txt
@@ -254,6 +254,7 @@ add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libCacheSim/profiler)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libCacheSim/utils)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libCacheSim/bin)
 add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libCacheSim/traceAnalyzer)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/libCacheSim/mrcProfiler)
 
 # #######################################
 # library install  #
